requirements.txt
beautifulsoup4==4.12.2
Cython==3.0.0
matplotlib==3.7.2
mlxtend==0.22.0
numpy==1.23.5
pandas==2.0.1
polars==0.19.8
pseudopeople==0.6.5
python_Levenshtein==0.23.0
PyYAML==6.0.1
lxml==4.9.3
